refactor(auth): absorb landing domain and split modals
- delete landing/ feature domain
- inline consent-frame content into consent-modal
- add views/components: login-modal, consent-modal, terms-modal, landing-grids, landing-modal
- landing-modal is a thin orchestrator composing the three modals
- add storybook stories for each modal component
