+++ src/features/auth/views/frames/index.ts
@@ -1,3 +1,2 @@
-export * from './callback-frame';
 export * from './auth-layout';
-export * from './consent-frame';
+export * from './callback-frame';
