+++ src/features/auth/views/index.ts
@@ -1,2 +1,2 @@
+export * from './components';
 export * from './frames';
-export * from './screens';
