--- src/features/landing/index.ts
@@ -1,1 +1,0 @@
-export * from './views';
