--- src/features/landing/views/index.ts
@@ -1,2 +1,0 @@
-export { AppRedirectPage } from './app-redirect';
-export { LandingModal } from './landing-modal';
